requirements.txt
Flask==1.0.2
google-cloud-storage== 1.7.0
google-cloud-speech==0.36.0
gunicorn==19.8.1
requests==2.19.1
WTForms-Appengine==0.1
line-bot-sdk==1.8.0
pydub==0.22.1
grpcio==1.15.0
